+++ e2e/nested-frame-navigation/app/shared/items-view-model.ts
@@ -1,0 +1,9 @@
+
+import { Observable } from "tns-core-modules/data/observable";
+import { Item } from "./item";
+
+export class ItemsViewModel extends Observable {
+    constructor(public items: Array<Item>) {
+        super();
+    }
+}